+++ tests/test_errors.py
@@ -3,7 +3,7 @@ import json
 import requests
 
 
-def test_decode_error(mocker, script_runner):
+def test_decode_error(mocker, script_runner, project):
     mock_response = mocker.Mock()
     mock_response.json.side_effect = json.decoder.JSONDecodeError("", "", 0)
     mock_get = mocker.patch("scrapyd_client.pyclient.requests.get", autospec=True)
@@ -14,7 +14,7 @@ def test_decode_error(mocker, script_runner):
     assert result.stdout.startswith("Received a malformed response:\n")
 
 
-def test_projects(mocker, script_runner):
+def test_projects(mocker, script_runner, project):
     mock_response = mocker.Mock()
     mock_response.json.return_value = {
         "status": "error",
@@ -28,7 +28,7 @@ def test_projects(mocker, script_runner):
     assert result.stdout == "Scrapyd responded with an error:\nSomething went wrong.\n"
 
 
-def test_connection_error(mocker, script_runner):
+def test_connection_error(mocker, script_runner, project):
     mock_get = mocker.patch("scrapyd_client.pyclient.requests.get", autospec=True)
     mock_get.side_effect = requests.ConnectionError()
     result = script_runner.run(["scrapyd-client", "projects"])